tidying up the admin script in the src folder since they are grown up in the last release
